- can finally use capturing lambda in ColorFade - Gamma class to encapsulate the gamma lut - '_s' literal operator for cleaner second->millisecond literals - namespace animator's current frame time global - namespace rrrus for function and hsv2rgb - default values for Animators private members instead of init in constructor - constexpr for some of the fixedp constructors so the compiler can render compile-time constants - frame delay adjustment hysteresis

+++ party_scarf_rrrus/rrrandom.h
@@ -24,12 +24,12 @@ static inline float randfRangeMaybe(float min, float max, float likelihood, floa
 	return min + (float)(random(1000001))/1000000.0f*scale;
 }
 
-static inline long randi(int32_t max)
+static inline long randi(long max)
 {
 	return random(max+1);
 }
 
-static inline long randiRange(int32_t min, int32_t max)
+static inline long randiRange(long min, long max)
 {
 	// return an int between min and max inclusive.
 	return min + random(max-min+1);
